+++ crates/common/common_utils/src/consts.rs
@@ -47,6 +47,10 @@ pub const X_FRM_CONNECTOR_NAME: &str = "x-frm-connector";
 pub const X_AUTHENTICATOR_CONNECTOR_NAME: &str = "x-auth-connector";
 /// Header key for merchant identification
 pub const X_MERCHANT_ID: &str = "x-merchant-id";
+/// Header key for payment method identification
+pub const X_PAYMENT_METHOD: &str = "x-payment-method";
+/// Header key for payment method type identification
+pub const X_PAYMENT_METHOD_TYPE: &str = "x-payment-method-type";
 /// Header key for reference identification
 pub const X_REFERENCE_ID: &str = "x-reference-id";
 /// Header key for resource identification