+++ crates/integrations/connector-integration/src/connectors.rs
@@ -260,6 +260,9 @@ pub mod juspay_upi_stack;
 pub mod axisbank;
 pub use self::axisbank::Axisbank;
 
+pub mod maya;
+pub use self::maya::Maya;
+
 pub mod tsys_transit;
 pub use self::tsys_transit::TsysTransit;
 pub mod twoc_twop_paco;